Week9/python/practice/audio_llm_project/requirements.txt
groq
pyaudio
python-dotenv
pyttsx3
